+++ modules/client/src/addresslistVoting/internal/client/methods.ts
@@ -51,6 +51,7 @@ import {
   EMPTY_PROPOSAL_METADATA_LINK,
   encodeProposalId,
   findLog,
+  getExtendedProposalId,
   InvalidAddressError,
   InvalidAddressOrEnsError,
   InvalidCidError,
@@ -71,14 +72,12 @@ import {
   resolveIpfsCid,
   SizeMismatchError,
   SortDirection,
-  SupportedNetwork,
-  SupportedNetworksArray,
   UNAVAILABLE_PROPOSAL_METADATA,
   UNSUPPORTED_PROPOSAL_METADATA_LINK,
   UnsupportedNetworkError,
 } from "@aragon/sdk-client-common";
 import { INSTALLATION_ABI, UPDATE_ABI } from "../constants";
-import { getExtendedProposalId } from "@aragon/sdk-client-common";
+import { ContractNames, getNetworkByAlias } from "@aragon/osx-commons-configs";
 
 /**
  * Methods module the SDK Address List Client
@@ -253,13 +252,15 @@ export class AddresslistVotingClientMethods extends ClientCore
     params: AddresslistVotingPluginPrepareInstallationParams,
   ): AsyncGenerator<PrepareInstallationStepValue> {
     const network = await this.web3.getProvider().getNetwork();
-    const networkName = network.name as SupportedNetwork;
-    if (!SupportedNetworksArray.includes(networkName)) {
-      throw new UnsupportedNetworkError(networkName);
+    const aragonNetwork = getNetworkByAlias(network.name);
+    if (!aragonNetwork) {
+      throw new UnsupportedNetworkError(network.name);
     }
     yield* prepareGenericInstallation(this.web3, {
       daoAddressOrEns: params.daoAddressOrEns,
-      pluginRepo: this.web3.getAddress("addresslistVotingRepoAddress"),
+      pluginRepo: this.web3.getAddress(
+        ContractNames.ADDRESSLIST_VOTING_REPO_PROXY,
+      ),
       version: params.versionTag,
       installationAbi: INSTALLATION_ABI,
       installationParams: [
@@ -267,7 +268,7 @@ export class AddresslistVotingClientMethods extends ClientCore
         params.settings.addresses,
       ],
       pluginSetupProcessorAddress: this.web3.getAddress(
-        "pluginSetupProcessorAddress",
+        ContractNames.PLUGIN_SETUP_PROCESSOR,
       ),
     });
   }
@@ -283,11 +284,13 @@ export class AddresslistVotingClientMethods extends ClientCore
   ): AsyncGenerator<PrepareUpdateStepValue> {
     yield* prepareGenericUpdate(this.web3, this.graphql, {
       ...params,
-      pluginRepo: this.web3.getAddress("addresslistVotingRepoAddress"),
+      pluginRepo: this.web3.getAddress(
+        ContractNames.ADDRESSLIST_VOTING_REPO_PROXY,
+      ),
       updateAbi: UPDATE_ABI[params.newVersion.build] ||
         params.updateAbi || [],
       pluginSetupProcessorAddress: this.web3.getAddress(
-        "pluginSetupProcessorAddress",
+        ContractNames.PLUGIN_SETUP_PROCESSOR,
       ),
     });
   }